Fundmental change to employee schedules
Check-in and check-out tables are now obsolete. Employee schedule table will now carry the times as strings instead of check-in check-out ids. Logic: Upon check-in will call a POST request to create new record with employee id, check-in time (from front-end) and log_datetime (auto generated by database). When employees check-out it will call a PUT request that will modify the POST from check-in with the check-out time. It will do this automatically under the assumption an employee does not check in and check out multiple times on the same date. WIP: PUT and GET not working yet; frontend will be after I get the calls working.